requirements.txt
# Core dependencies
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Image processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.2

# PDF processing
pypdf==3.17.1
PyMuPDF==1.23.8
pdf2image==1.16.3
pdfplumber==0.10.3

# Google Gemini
google-generativeai==0.3.2

# CLI
typer==0.9.0
rich==13.7.0

# Telegram
python-telegram-bot==20.7

# Web
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6

# Email
imapclient==2.3.1

# Excel export
openpyxl==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
httpx==0.25.2
